move id image button
